requirements.txt
asgiref==3.8.1
Django==5.1.6
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
djangorestframework==3.15.2
phonenumbers==8.13.55
python-dotenv==1.0.1
sqlparse==0.5.3
